+++ src/commands/up/index.ts
@@ -5,6 +5,18 @@ import { UpOptions } from '../types';
 const backendSuccessIndicator = 'Running on http://localhost:8000';
 const frontendSuccessIndicator = 'ready - started server on 0.0.0.0:3000';
 
+function startNetwork () {
+  try {
+    const commands = [
+      'docker network rm ops-console 2> /dev/null',
+      'docker network create -d bridge ops-console 2> /dev/null'
+    ].join('\n');
+    runCommand(commands);
+  } catch (e) {
+    logger.error(`Error launching ops console network: ${e}`);
+  }
+}
+
 function runBackend (tag: string) {
   try {
     const commands = [
@@ -12,7 +24,7 @@ function runBackend (tag: string) {
       `docker pull 849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-api:latest-${tag}`,
       'docker container stop ops-api || true',
       'docker container rm ops-api || true',
-      `docker run --name ops-api -v $HOME/.aws:/root/.aws -v $(pwd):/config --env CONFIG_PATH="../config/example.yml" -i -p 8000:8000 "849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-api:latest-${tag}";`
+      `docker run --name ops-api -v $HOME/.aws:/root/.aws -v $(pwd):/config --env CONFIG_PATH="../config/example.yml" -i -p 8000:8000 --network=ops-console "849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-api:latest-${tag}";`
     ].join(';\n');
     const childProcess = runCommand(commands);
     childProcess.stdout.on('data', (data) => {
@@ -21,7 +33,7 @@ function runBackend (tag: string) {
       }
     });
   } catch (e) {
-    logger.error(`Error launching ops console servers: ${e}`);
+    logger.error(`Error launching ops console backend: ${e}`);
   }
 }
 
@@ -32,7 +44,7 @@ function runFrontend (tag: string) {
       `docker pull 849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-frontend:latest-${tag}`,
       'docker container stop ops-frontend || true',
       'docker container rm ops-frontend || true',
-      `docker run --name ops-frontend -v $HOME/.aws:/root/.aws -i -p 3000:3000 "849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-frontend:latest-${tag}";`
+      `docker run --name ops-frontend -i -p 3000:3000 --network=ops-console "849087520365.dkr.ecr.us-east-1.amazonaws.com/ops-frontend:latest-${tag}";`
     ].join(';\n');
     const childProcess = runCommand(commands);
     childProcess.stdout.on('data', (data) => {
@@ -41,11 +53,12 @@ function runFrontend (tag: string) {
       }
     });
   } catch (e) {
-    logger.error(`Error launching ops console servers: ${e}`);
+    logger.error(`Error launching ops console frontend: ${e}`);
   }
 }
 
 async function up (options: UpOptions) {
+  startNetwork();
   const { arm } = options;
   const tag = arm ? 'arm' : 'x86';
   runBackend(tag);